update strava to latest API

+++ fivekmrun_app_flutter/lib/common/double_extensions.dart
@@ -1,4 +1,4 @@
-extension DoubleExtension on double {
+extension DoubleExtension on int {
   String metersToKm() {
     return (this / 1000).toStringAsFixed(2);
   }
